RelScanner

+++ src/wrapper.cpp
@@ -27,6 +27,14 @@ using namespace jjjj222;
 //------------------------------------------------------------------------------
 //   
 //------------------------------------------------------------------------------
+string dump_field_type_str(const FIELD_TYPE&);
+string dump_tiny_type_str(const DataType&);
+DataType field_to_data_type(const FIELD_TYPE&);
+FIELD_TYPE tiny_to_field_type(const DataType&);
+
+//------------------------------------------------------------------------------
+//   
+//------------------------------------------------------------------------------
 string dump_field_type_str(const FIELD_TYPE& type)
 {
     if (type == INT) {
@@ -252,6 +260,12 @@ string TinySchema::dump_str() const
 //------------------------------------------------------------------------------
 //   TinyTuple
 //------------------------------------------------------------------------------
+TinyTuple::TinyTuple(TinyRelation* r)
+: _tuple(NULL)
+{
+    _tuple = new Tuple(r->create_null_tuple());
+}
+
 TinyTuple::TinyTuple(const Tuple& tuple)
 : _tuple(NULL)
 {
@@ -319,9 +333,19 @@ void TinyTuple::init()
     }
 }
 
+void TinyTuple::set_null()
+{
+    _tuple->null();
+}
+
 TinySchema TinyTuple::get_tiny_schema() const
 {
     return TinySchema(_tuple->getSchema());
+}
+
+size_t TinyTuple::size() const
+{
+    return get_tiny_schema().size();
 }
 
 DataValue TinyTuple::get_value(const string& column_name) const
@@ -404,6 +428,9 @@ void TinyTuple::assign(const TinyTuple& rhs)
 
 string TinyTuple::dump_str() const
 {
+    if (is_null())
+        return "<DELETE>";
+
     vector<pair<string, FIELD_TYPE>> name_type_list = get_tiny_schema().get_name_type_list();
 
     string tmp = "(";
@@ -455,6 +482,11 @@ void TinyTuple::dump() const
   //union Field getField(int offset) const; // returns INT_MIN if out of bound
 }
 
+vector<string> TinyTuple::dump_str_list() const
+{
+    return str_list();
+}
+
 //------------------------------------------------------------------------------
 //   
 //------------------------------------------------------------------------------
@@ -481,6 +513,49 @@ vector<TinyTuple> TinyBlock::get_tiny_tuples() const
     return tmp;
 }
 
+TinyTuple TinyBlock::get_tiny_tuple(size_t i) const
+{
+//    //assert(
+    return _block->getTuple(i);
+}
+
+size_t TinyBlock::size()
+{
+    return get_tuples().size();
+}
+
+void TinyBlock::push_back(const TinyTuple& tuple)
+{
+    _block->appendTuple(tuple);
+}
+
+void TinyBlock::set_tuple(size_t i, const TinyTuple& tuple)
+{
+    _block->setTuple(i, tuple);
+}
+
+void TinyBlock::remove_null_tuple()
+{
+    vector<Tuple> tuples = get_tuples();
+    _block->clear();
+    for (const auto& tuple : tuples) {
+        if (!tuple.isNull()) {
+            push_back(tuple);
+            //dump_normal(TinyTuple(tuple));
+        }
+    }
+}
+
+void TinyBlock::null_tuple(size_t i)
+{
+    _block->nullTuple(i);
+}
+
+void TinyBlock::clear()
+{
+    _block->clear();
+}
+
 bool TinyBlock::empty() const
 {
     vector<Tuple> tuples = get_tuples();
@@ -489,6 +564,374 @@ bool TinyBlock::empty() const
             return false;
     }
     return true;
+}
+
+void TinyBlock::dump() const
+{
+    vector<TinyTuple> tuples = get_tiny_tuples();
+    for (const auto& tuple : tuples) {
+        tuple.dump();
+    }
+}
+
+string TinyBlock::dump_str() const
+{
+    return "TODO";
+}
+
+vector<string> TinyBlock::dump_str_list() const
+{
+    size_t count = 0;
+    vector<TinyTuple> tuples = get_tiny_tuples();
+
+    vector<string> res;
+    for (const auto& tuple : tuples) {
+        if (tuple.is_null()) {
+            size_t num = tuple.size();
+            count += num;
+            for (size_t i = 0; i < num; ++i) {
+                res.push_back(""); 
+            }
+        } else {
+            const vector<string> str_list = tuple.dump_str_list();
+            count += str_list.size();
+            add_into(res, str_list); 
+        }
+        //tuple.dump();
+    }
+
+    for (size_t i = count; i < HwMgr::ins()->get_block_size(); ++i) {
+        res.push_back("");
+    }
+    return res;
+}
+
+//------------------------------------------------------------------------------
+//   
+//------------------------------------------------------------------------------
+MemIter::MemIter(TinyRelation* r, size_t pos)
+: _relation(r)
+, _pos(pos)
+{
+    ;
+}
+
+//void MemIter::clear_block()
+//{
+//    size_t block_idx = get_block_idx();
+//    TinyBlock block = HwMgr::ins()->get_mem_block(block_idx);
+//    block.clear();
+//}
+
+//void MemIter::push_back(const TinyTuple& tuple)
+//{
+//    size_t block_idx = get_block_idx();
+//    TinyBlock block = HwMgr::ins()->get_mem_block(block_idx);
+//    block.push_back(tuple);
+//}
+void MemIter::set_tuple(const TinyTuple& tuple)
+{
+    size_t block_idx = get_block_idx();
+    size_t tuple_idx = get_tuple_idx();
+    TinyBlock block = HwMgr::ins()->get_mem_block(block_idx);
+    block.set_tuple(tuple_idx, tuple);
+}
+
+TinyTuple MemIter::get_tuple() const
+{
+    size_t block_idx = get_block_idx();
+    size_t tuple_idx = get_tuple_idx();
+    TinyBlock block = HwMgr::ins()->get_mem_block(block_idx);
+    return block.get_tiny_tuple(tuple_idx);
+}
+
+size_t MemIter::get_block_idx() const
+{
+    return _pos / _relation->tuple_per_block();
+}
+
+size_t MemIter::get_tuple_idx() const
+{
+    return _relation->get_tuple_idx_by_pos(_pos);
+}
+
+bool MemIter::is_equal_to(const MemIter& rhs) const
+{
+    assert(_relation == rhs._relation);
+
+    bool res = _pos == rhs._pos;
+    return res;
+}
+
+string MemIter::dump_str() const
+{
+    //return tiny_dump_str(_pos);
+    return tiny_dump_str(get_block_idx(), get_tuple_idx());
+}
+//------------------------------------------------------------------------------
+//   RelIter
+//------------------------------------------------------------------------------
+RelIter::RelIter(TinyRelation* r)
+: _relation(r)
+//, _block(0)
+, _pos(0)
+//, _mem_idx(0)
+{
+    ;
+}
+
+RelIter::RelIter(TinyRelation* r, size_t pos)
+: _relation(r)
+, _pos(pos)
+{
+    ;
+}
+//void RelIter::inc()
+//{
+//    // TODO
+//}
+//
+bool RelIter::is_equal_to(const RelIter& rhs) const
+{
+    assert(_relation == rhs._relation);
+    //assert(_mem_idx == rhs._mem_idx);
+
+    //bool res = _block == rhs._block && _pos == rhs._pos;
+    bool res = _pos == rhs._pos;
+    return res;
+}
+
+size_t RelIter::get_block_idx() const
+{
+    return _relation->get_block_idx_by_pos(_pos);
+}
+
+size_t RelIter::get_tuple_idx() const
+{
+    return _relation->get_tuple_idx_by_pos(_pos);
+}
+
+TinyTuple RelIter::load_to_mem(size_t mem_idx) const
+{
+    pair<size_t, size_t> block_tuple_idx = _relation->get_idx_by_pos(_pos);
+    size_t block_idx = block_tuple_idx.first;
+    size_t tuple_idx = block_tuple_idx.second;
+    _relation->load_block_to_mem(block_idx, mem_idx);
+    TinyBlock block = HwMgr::ins()->get_mem_block(mem_idx);
+    return  block.get_tiny_tuple(tuple_idx);
+}
+
+TinyTuple RelIter::get_from_mem(size_t mem_idx) const
+{
+    pair<size_t, size_t> block_tuple_idx = _relation->get_idx_by_pos(_pos);
+    size_t tuple_idx = block_tuple_idx.second;
+    TinyBlock block = HwMgr::ins()->get_mem_block(mem_idx);
+    return  block.get_tiny_tuple(tuple_idx);
+}
+
+void RelIter::skip_null()
+{
+    while (!is_end() && is_null()) {
+        ++_pos;
+    }
+}
+
+bool RelIter::is_null() const
+{
+    return _relation->is_null(_pos);
+}
+
+bool RelIter::is_end() const
+{
+    return _relation->end() == *this;
+}
+
+string RelIter::dump_str() const
+{
+    //return tiny_dump_str(_pos);
+    return tiny_dump_str(get_block_idx(), get_tuple_idx());
+}
+//TinyTuple RelIter::get_tuple() const
+//{
+//    _relation->load_block_to_mem(_block, _mem_idx);
+//    //TinyBlock block = HwMgr::ins()->get_tiny_block(_mem_idx);
+//    TinyBlock block = HwMgr::ins()->get_mem_block(_mem_idx);
+//    vector<TinyTuple> tuples = block.get_tiny_tuples();
+//
+//    if (_pos >= tuples.size()) {
+//        return _relation->create_null_tuple();
+//    }
+//
+//    return tuples[_pos];
+//    //return block.get_tiny_tuple(_pos);
+//
+//            //_relation->getBlock(i, mem_index);
+//            //Block* block = HwMgr::ins()->get_mem_block(mem_index);
+//            //vector<Tuple> tuples = block->getTuples();
+//            //for (const auto& tuple : tuples) {
+//            //    //dump_normal(TinyTuple(tuple));
+//            //    if (!tuple.isNull())
+//            //        table.add_row(TinyTuple(tuple).str_list());
+//            //}
+//        //}
+//}
+
+//------------------------------------------------------------------------------
+//   
+//------------------------------------------------------------------------------
+RelScanner::RelScanner(TinyRelation* r, size_t base_idx, size_t mem_size)
+: _relation(r)
+, _base_idx(base_idx)
+, _mem_size(mem_size)
+, _m_iter(r, 0)
+, _m_iter_end(r, 0)
+, _iter(r->begin())
+{
+    assert(_relation != NULL);
+    assert(_base_idx >= 0 && _base_idx < HwMgr::ins()->get_mem_size());
+    assert(_mem_size >= 1);
+    assert(_base_idx + _mem_size - 1 < HwMgr::ins()->get_mem_size());
+}
+
+
+TinyTuple RelScanner::get_next()
+{
+    //move_to_non_null();
+    _iter.skip_null();
+    if (_iter.is_end()) {
+        return _relation->create_null_tuple();
+    }
+
+    if (_m_iter == _m_iter_end) {
+        load_to_mem();
+    }
+
+    TinyTuple res = _m_iter.get_tuple();
+    ++_m_iter;
+
+    return res;
+}
+
+void RelScanner::clear_mem()
+{
+    for (size_t i = 0; i < _mem_size; ++i) {
+        size_t mem_idx = _base_idx + i;
+        clear_mem_block(mem_idx);
+    }
+}
+
+void RelScanner::clear_mem_block(size_t i)
+{
+    assert(i >= _base_idx);
+    assert(i < _base_idx + _mem_size);
+
+    HwMgr::ins()->get_mem_block(i).clear();
+}
+
+//void RelScanner::move_to_non_null()
+//{
+//    while (!_iter.is_end() && _iter.is_null()) {
+//        ++_iter;
+//    }
+//}
+
+void RelScanner::load_to_mem()
+{
+    clear_mem();
+
+    _m_iter = m_begin();
+    bool reuse_block = false;
+    size_t old_block_idx = 0;
+    size_t load_idx = get_last_mem_block();
+    for (_m_iter = m_begin(); _m_iter != m_load_end(); ++_m_iter) {
+        //move_to_non_null();
+        _iter.skip_null();
+        if (_iter.is_end()) {
+            break;
+        }
+
+        size_t block_idx = _iter.get_block_idx();
+        TinyTuple tuple(_relation);
+        if (reuse_block && block_idx == old_block_idx) {
+            tuple = _iter.get_from_mem(load_idx); 
+        } else {
+            reuse_block = true;
+            old_block_idx = block_idx;
+            tuple = _iter.load_to_mem(load_idx); 
+        }
+        ++_iter;
+
+        _m_iter.set_tuple(tuple);
+    }
+    //dump_normal(_iter);
+
+    //_m_iter_end = _m_iter;
+    if (_m_iter == m_load_end()) {
+        //move_to_non_null();
+        _iter.skip_null();
+        if (!_iter.is_end()) {
+            size_t last_block_idx = _iter.get_block_idx();
+            while (!_iter.is_end() && _iter.get_block_idx() == last_block_idx) {
+                ++_iter;
+            }
+            //dump_normal(last_block_idx);
+            _relation->load_block_to_mem(last_block_idx, load_idx);
+            TinyBlock block = HwMgr::ins()->get_mem_block(load_idx);
+            block.remove_null_tuple();
+            //block.dump();
+            for (size_t i = 0; i < block.size(); ++i) {
+                ++_m_iter;
+            }
+            //dump_normal(_m_iter);
+        }
+    }
+
+    _m_iter_end = _m_iter;
+    _m_iter = m_begin();
+    //dump_normal(_m_iter);
+    //dump_normal(_m_iter_end);
+    //dump_normal(_iter);
+    //_m_iter = m_begin();
+    //move_to_non_null();
+
+
+    //TinyTuple res = _iter.load_to_mem(_mem_idx); 
+
+    //++_iter;
+    //return res;
+}
+
+//TinyTuple RelScanner::get_from_mem()
+//{
+//}
+
+bool RelScanner::is_end() const
+{
+    bool res = _iter.is_end() && _m_iter == _m_iter_end;
+    return res;
+}
+
+MemIter RelScanner::m_begin()
+{
+    size_t mem_pos = _base_idx * _relation->tuple_per_block();
+    return MemIter(_relation, mem_pos);
+}
+
+MemIter RelScanner::m_load_end()
+{
+    size_t mem_pos = (_base_idx + _mem_size - 1) * _relation->tuple_per_block();
+    return MemIter(_relation, mem_pos);
+}
+
+MemIter RelScanner::m_end()
+{
+    size_t mem_pos = (_base_idx + _mem_size) * _relation->tuple_per_block();
+    return MemIter(_relation, mem_pos);
+}
+
+size_t RelScanner::get_last_mem_block() const
+{
+    return _base_idx + _mem_size - 1;
 }
 
 //------------------------------------------------------------------------------
@@ -506,14 +949,18 @@ void TinyRelation::push_back(const TinyTuple& tuple)
 {
     //dump_pretty(tuple);
     size_t mem_index = 0;
-    Block* block = HwMgr::ins()->get_mem_block(mem_index);
+    //Block* block = HwMgr::ins()->get_mem_block(mem_index);
+    TinyBlock block = HwMgr::ins()->get_mem_block(mem_index);
     if (next_is_new_block()) {
-        block->clear();
-        block->appendTuple(tuple);
+        //block->clear();
+        block.clear();
+        //block->appendTuple(tuple);
+        block.push_back(tuple);
         _relation->setBlock(_relation->getNumOfBlocks(), mem_index);
     } else {
         _relation->getBlock(_relation->getNumOfBlocks()-1, mem_index);
-        block->appendTuple(tuple); // append the tuple
+        //block->appendTuple(tuple); // append the tuple
+        block.push_back(tuple);
         _relation->setBlock(_relation->getNumOfBlocks()-1, mem_index);
     }
     //relation_ptr->setBlock(relation_ptr->getNumOfBlocks(),memory_block_index);
@@ -573,6 +1020,8 @@ void TinyRelation::refresh_block_num()
         if (_space[i] == total_size - 1) {
             total_size--;
             _space.pop_back();
+        } else {
+            break;
         }
     }
 
@@ -594,16 +1043,20 @@ bool TinyRelation::save_block_to_disk(size_t block_index, size_t mem_index) cons
     return _relation->setBlock(block_index, mem_index);
 }
 
-bool TinyRelation::reduce_blocks_to(size_t block_num) const
+void TinyRelation::reduce_blocks_to(size_t block_num) const
 {
-    //dump_normal(block_num);
-    return _relation->deleteBlocks(block_num);
+    if (block_num < get_num_of_block()) {
+        _relation->deleteBlocks(block_num);
+    }
 }
 
-TinyTuple TinyRelation::create_tuple() const
+TinyTuple TinyRelation::create_tuple(bool is_null) const
 {
     TinyTuple t = _relation->createTuple();
     t.init();
+    if (is_null) {
+        t.set_null();
+    }
     return t;
 }
 
@@ -645,6 +1098,45 @@ size_t TinyRelation::get_num_of_block() const
 size_t TinyRelation::tuple_per_block() const
 {
     return get_tiny_schema().tuple_per_block();
+}
+
+size_t TinyRelation::get_block_idx_by_pos(size_t pos) const
+{
+    return pos / tuple_per_block();
+}
+
+size_t TinyRelation::get_tuple_idx_by_pos(size_t pos) const
+{
+    return pos % tuple_per_block();
+}
+
+pair<size_t, size_t> TinyRelation::get_idx_by_pos(size_t pos) const
+{
+    pair<size_t, size_t> tmp;
+    tmp.first = pos / tuple_per_block();
+    tmp.second = pos % tuple_per_block();
+    return tmp;
+}
+
+bool TinyRelation::is_null(size_t pos) const
+{
+    return is_contain(_space, pos);
+}
+
+TinyRelation::iterator TinyRelation::begin()
+{
+    return iterator(this);
+}
+
+TinyRelation::iterator TinyRelation::end()
+{
+    size_t end_pos = get_total_pos();
+    return iterator(this, end_pos);
+}
+
+size_t TinyRelation::get_total_pos() const
+{
+    return size() + _space.size();
 }
 
 bool TinyRelation::next_is_new_block() const
